pyproject.toml
[tool.poetry]
name = "book-reader-backend"
version = "0.1.0"
description = ""
authors = ["andriiskrypnyk <andre.beltukov5656@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.1"
djangorestframework = "^3.14.0"
celery = "^5.3.6"
drf-yasg = "^1.21.7"
python-dotenv = "^1.0.0"
setuptools = "^69.0.3"
pytest = "^7.4.4"
pytest-django = "^4.7.0"
django-configurations = "^2.5"
psycopg2-binary = "^2.9.9"
djangorestframework-jwt = "^1.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
